refactor: watch 도메인 클래스 이관
