+++ src/main/java/com/ssginc/showpingrefactoring/domain/watch/service/WatchService.java
@@ -1,0 +1,15 @@
+package com.ssginc.showpingrefactoring.domain.watch.service;
+
+import com.ssginc.showpingrefactoring.domain.watch.dto.request.WatchRequestDto;
+import com.ssginc.showpingrefactoring.domain.watch.dto.resopnse.WatchResponseDto;
+import com.ssginc.showpingrefactoring.domain.watch.entity.Watch;
+
+import java.util.List;
+
+public interface WatchService {
+
+    List<WatchResponseDto> getWatchHistoryByMemberNo(Long memberNo);
+
+    Watch insertWatchHistory(WatchRequestDto watchRequestDto, Long memberNo);
+
+}
