--- src/main/java/com/ssginc/showpingrefactoring/vod/service/WatchService.java
@@ -1,14 +1,0 @@
-package com.ssginc.showpingrefactoring.vod.service;
-
-import com.ssginc.showpingrefactoring.vod.dto.request.WatchRequestDto;
-import com.ssginc.showpingrefactoring.vod.dto.response.WatchResponseDto;
-
-import java.util.List;
-
-public interface WatchService {
-
-    List<WatchResponseDto> getWatchHistoryByMemberNo(Long memberNo);
-
-    Watch insertWatchHistory(WatchRequestDto watchRequestDto, Long memberNo);
-
-}
